+++ src/Pages/About.js
@@ -3,39 +3,8 @@ import "../App.css";
 import NavBar from "../Components/NavBar";
 import { Helmet } from "react-helmet";
 import Form from '../Components/Form';
-import axios from "axios";
 
 const About = () => {
-
-  const submitSighting = sighting => {
-    const newObj = {
-      witness: sighting.name,
-      seenDate: sighting.date,
-      location: sighting.location,
-      description: sighting.desc,
-      imageUrl: sighting.uploadedImg
-    };
-
-    const axiosConfig = {
-      headers: {
-        "Content-Type": "application/json;charset=UTF-8",
-        "Access-Control-Allow-Origin": "*",
-        Authorization: this.state.jwt
-      }
-    };
-    axios
-      .post(
-        "https://mothman-server.herokuapp.com/users/new-sighting",
-        newObj,
-        axiosConfig
-      )
-      .then(response => {
-        console.log(response);
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  };
 
   return (
     <div
@@ -81,7 +50,7 @@ const About = () => {
               Moth Man project, your story will become available for the rest
               of the web to see.
             </p>
-            <Form appSubmitSighting={submitSighting} />
+            <Form />
 
             <div style={{ height: 200 }} />
             <p className="lead mb-0">The Moth Maps Project Team</p>
